// src/Type/translationType/forgetPasswordTranslationType.type.ts
type forgetInterFace = {
    passwordRecovery: string;
    forgetPasswordText: string;
    close: string;
    send: string;
    forgetPasswordButton: string;
    email: string;
    userName: string;

}

export type forgetPasswordTranslationType = {
    locale: string;
    Translation: forgetInterFace;
}